refactor: [#521]: reorganize new config modules
