+++ packages/configuration/src/lib.rs
@@ -229,10 +229,7 @@
 //! [health_check_api]
 //! bind_address = "127.0.0.1:1313"
 //!```
-pub mod health_check_api;
-pub mod http_tracker;
-pub mod tracker_api;
-pub mod udp_tracker;
+pub mod sections;
 
 use std::collections::HashMap;
 use std::net::IpAddr;
