+++ packages/configuration/src/sections/tracker_api.rs
@@ -1,6 +1,6 @@
-//! Validated configuration for the HTTP Tracker service.
+//! Validated configuration for the Tracker API service.
 //!
-//! [``crate::HttpTracker``] is a DTO containing the parsed data from the toml
+//! [``crate::HttpApi``] is a DTO containing the parsed data from the toml
 //! file.
 //!
 //! This configuration is a first level of validation that can be perform
@@ -14,7 +14,7 @@ use std::net::SocketAddr;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-use crate::HttpTracker;
+use crate::{AccessTokens, HttpApi};
 
 /// Errors that can occur when validating the plain configuration.
 #[derive(Error, Debug, PartialEq)]
@@ -38,6 +38,7 @@ pub struct Config {
     ssl_enabled: bool,
     ssl_cert_path: Option<String>, // todo: use Path
     ssl_key_path: Option<String>,  // todo: use Path
+    access_tokens: AccessTokens,
 }
 
 impl Config {
@@ -47,10 +48,10 @@ impl Config {
     }
 }
 
-impl TryFrom<HttpTracker> for Config {
+impl TryFrom<HttpApi> for Config {
     type Error = ValidationError;
 
-    fn try_from(config: HttpTracker) -> Result<Self, Self::Error> {
+    fn try_from(config: HttpApi) -> Result<Self, Self::Error> {
         let socket_addr = match config.bind_address.parse::<SocketAddr>() {
             Ok(socket_addr) => socket_addr,
             Err(_err) => {
@@ -90,11 +91,12 @@ impl TryFrom<HttpTracker> for Config {
             ssl_enabled: config.ssl_enabled,
             ssl_cert_path: config.ssl_cert_path,
             ssl_key_path: config.ssl_key_path,
+            access_tokens: config.access_tokens,
         })
     }
 }
 
-impl From<Config> for HttpTracker {
+impl From<Config> for HttpApi {
     fn from(config: Config) -> Self {
         Self {
             enabled: config.enabled,
@@ -102,22 +104,27 @@ impl From<Config> for HttpTracker {
             ssl_enabled: config.ssl_enabled,
             ssl_cert_path: config.ssl_cert_path,
             ssl_key_path: config.ssl_key_path,
+            access_tokens: config.access_tokens,
         }
     }
 }
 
 #[cfg(test)]
 mod tests {
+
+    use std::collections::HashMap;
+
     use super::*;
 
     #[test]
     fn it_should_return_an_error_when_the_bind_address_is_not_a_valid_socket_address() {
-        let plain_config = HttpTracker {
+        let plain_config = HttpApi {
             enabled: true,
             bind_address: "300.300.300.300:7070".to_string(),
             ssl_enabled: true,
             ssl_cert_path: None,
             ssl_key_path: Some("./localhost.key".to_string()),
+            access_tokens: HashMap::new(),
         };
 
         assert_eq!(
@@ -129,17 +136,20 @@ mod tests {
     }
 
     mod when_ssl_is_enabled {
-        use crate::http_tracker::{Config, ValidationError};
-        use crate::HttpTracker;
+        use std::collections::HashMap;
+
+        use crate::sections::tracker_api::{Config, ValidationError};
+        use crate::HttpApi;
 
         #[test]
         fn it_should_return_an_error_when_ssl_is_enabled_but_the_cert_path_is_not_provided() {
-            let plain_config = HttpTracker {
+            let plain_config = HttpApi {
                 enabled: true,
-                bind_address: "127.0.0.1:7070".to_string(),
+                bind_address: "127.0.0.1:1212".to_string(),
                 ssl_enabled: true,
                 ssl_cert_path: None,
                 ssl_key_path: Some("./localhost.key".to_string()),
+                access_tokens: HashMap::new(),
             };
 
             assert_eq!(Config::try_from(plain_config), Err(ValidationError::MissingSslCertPath));
@@ -147,12 +157,13 @@ mod tests {
 
         #[test]
         fn it_should_return_an_error_when_ssl_is_enabled_but_the_cert_path_is_empty() {
-            let plain_config = HttpTracker {
+            let plain_config = HttpApi {
                 enabled: true,
-                bind_address: "127.0.0.1:7070".to_string(),
+                bind_address: "127.0.0.1:1212".to_string(),
                 ssl_enabled: true,
                 ssl_cert_path: Some(String::new()),
                 ssl_key_path: Some("./localhost.key".to_string()),
+                access_tokens: HashMap::new(),
             };
 
             assert_eq!(Config::try_from(plain_config), Err(ValidationError::MissingSslCertPath));
@@ -160,12 +171,13 @@ mod tests {
 
         #[test]
         fn it_should_return_an_error_when_ssl_is_enabled_but_the_key_path_is_not_provided() {
-            let plain_config = HttpTracker {
+            let plain_config = HttpApi {
                 enabled: true,
-                bind_address: "127.0.0.1:7070".to_string(),
+                bind_address: "127.0.0.1:1212".to_string(),
                 ssl_enabled: true,
                 ssl_cert_path: Some("./localhost.crt".to_string()),
                 ssl_key_path: None,
+                access_tokens: HashMap::new(),
             };
 
             assert_eq!(Config::try_from(plain_config), Err(ValidationError::MissingSslKeyPath));
@@ -173,12 +185,13 @@ mod tests {
 
         #[test]
         fn it_should_return_an_error_when_ssl_is_enabled_but_the_key_path_is_empty() {
-            let plain_config = HttpTracker {
+            let plain_config = HttpApi {
                 enabled: true,
-                bind_address: "127.0.0.1:7070".to_string(),
+                bind_address: "127.0.0.1:1212".to_string(),
                 ssl_enabled: true,
                 ssl_cert_path: Some("./localhost.crt".to_string()),
                 ssl_key_path: Some(String::new()),
+                access_tokens: HashMap::new(),
             };
 
             assert_eq!(Config::try_from(plain_config), Err(ValidationError::MissingSslKeyPath));
